requirements.txt
absl-py==1.4.0
accelerate==1.3.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.6.0
anyio==4.2.0
app==0.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bert-serving-client==1.10.0
bert-serving-server==1.10.0
binaryornot==0.4.4
black==24.10.0
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
comtypes==1.3.1
contourpy==1.2.0
cookiecutter==2.6.0
coverage==7.6.3
cycler==0.12.1
debugpy==1.6.8
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
dlib==19.24.2
dnspython==2.7.0
email_validator==2.2.0
environs==11.0.0
executing==1.2.0
factory_boy==3.3.1
Faker==30.6.0
fastapi==0.110.0
fastjsonschema==2.18.0
filelock==3.13.1
flake8==7.1.1
flake8-blind-except==0.2.1
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-isort==6.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Compress==1.14
Flask-Cors==4.0.0
Flask-DebugToolbar==0.16.0
Flask-JSON==0.4.0
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Static-Digest==0.4.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.4.0
gevent==24.10.2
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.56.2
gTTS==2.5.4
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.21.1
idna==3.10
importlib_metadata==7.1.0
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.14.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==2.15.0
keras-utils==1.0.13
kiwisolver==1.4.5
libclang==16.0.6
limits==4.0.1
Mako==1.3.5
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
mpremote==1.22.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.5.7
networkx==3.2.1
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
ordered-set==4.1.0
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pep8-naming==0.14.1
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
playsound==1.3.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydocstyle==6.3.0
pyflakes==3.2.0
pygame==2.5.2
Pygments==2.15.1
pyodbc==5.2.0
pyparsing==3.1.1
pypiwin32==223
pyproject-toml==0.0.10
pyserial==3.5
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-json-logger==2.0.7
python-slugify==8.0.4
python-vlc==3.0.20123
pyttsx3==2.90
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.34.0
regex==2023.12.25
replicate==1.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.5.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
SpeechRecognition==3.10.1
SQLAlchemy==2.0.36
stack-data==0.6.2
starlette==0.36.3
supervisor==4.2.5
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.18.0.post0
tflearn==0.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
torch==2.2.1
torchvision==0.17.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.39.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.1.30
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.27.1
virtualenv==20.28.1
waitress==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.8
websocket-client==1.6.1
WebTest==3.0.1
Werkzeug==3.0.4
wrapt==1.14.1
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==7.1.0
